requirements.txt
# requirements.txt  •  minimal but complete

# ─── AI models ────────────────────────────────────────────────────────────────
google-generativeai==0.8.5       # Imagen / Gemini
anthropic==0.49.0                # Claude text generation

# ─── Google APIs (Search Console, auth) ───────────────────────────────────────
google-api-python-client==2.166.0
google-auth==2.38.0              # indirect for the line above

# ─── Networking / env / images ───────────────────────────────────────────────
requests==2.32.3
python-dotenv==1.1.0
Pillow==11.2.1                   # convert raw bytes to images for WP upload

# ─── Data wrangling & DB access ──────────────────────────────────────────────
pandas==2.2.3
numpy==2.2.4
mysql-connector-python==9.2.0

# ─── (Optional) Flask micro‑API to trigger jobs ──────────────────────────────
flask==3.1.0
gunicorn==23.0.0

# ─── Google APIs ──────────────────────────────────────────────────────────────
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1        # ← NEW (for InstalledAppFlow)

nltk==3.9.1
aiohttp==3.11.6
